Move plugins in folders to root
