+++ packages/create-next-stack/src/main/plugins/chakra-ui.ts
@@ -1,5 +1,5 @@
 import endent from "endent"
-import { Plugin } from "../../plugin"
+import { Plugin } from "../plugin"
 
 export const chakraUIPlugin: Plugin = {
   id: "chakra-ui",
