+++ packages/create-next-stack/src/main/setup/setup.ts
@@ -6,17 +6,17 @@ import { inDebugMode } from "../helpers/in-debug-mode"
 import { time } from "../helpers/time"
 import { logDebug, logInfo } from "../logging"
 import { evalOptionalProperty, evalProperty, Plugin } from "../plugin"
-import { chakraUIPlugin } from "../plugins/chakra-ui/chakra-ui"
+import { chakraUIPlugin } from "../plugins/chakra-ui"
 import { createNextStackPlugin } from "../plugins/create-next-stack/create-next-stack"
-import { cssModulesPlugin } from "../plugins/css-modules/css-modules"
+import { cssModulesPlugin } from "../plugins/css-modules"
 import { emotionPlugin } from "../plugins/emotion"
 import { eslintPlugin } from "../plugins/eslint"
 import { formattingPreCommitHookPlugin } from "../plugins/formatting-pre-commit-hook"
 import { formikPlugin } from "../plugins/formik"
 import { framerMotionPlugin } from "../plugins/framer-motion"
 import { githubActionsPlugin } from "../plugins/github-actions"
-import { mantinePlugin } from "../plugins/mantine/mantine"
-import { materialUIPlugin } from "../plugins/material-ui/material-ui"
+import { mantinePlugin } from "../plugins/mantine"
+import { materialUIPlugin } from "../plugins/material-ui"
 import { netlifyPlugin } from "../plugins/netlify"
 import { nextPlugin } from "../plugins/next"
 import { npmPlugin } from "../plugins/npm"
@@ -27,7 +27,7 @@ import { reactPlugin } from "../plugins/react"
 import { reactHookFormPlugin } from "../plugins/react-hook-form"
 import { reactIconsPlugin } from "../plugins/react-icons"
 import { reactQueryPlugin } from "../plugins/react-query"
-import { sassPlugin } from "../plugins/sass/sass"
+import { sassPlugin } from "../plugins/sass"
 import { styledComponentsPlugin } from "../plugins/styled-components"
 import { tailwindCSSPlugin } from "../plugins/tailwind-css"
 import { typescriptPlugin } from "../plugins/typescript"
